Ported some changes from the old emacs repository.
Some changes which were comitted on the old emacs repository could not be pushed
to said repo, as it was gone.

As history has been re-written for the new repo, a simple merge would not have
had the desired effect.  Also rebasing didn’t quite work the way I expected it
to.  So I’m just applying one commit here which contains the changes I made.